Update replicanti.js
Man I was rushed

+++ src/core/replicanti.js
@@ -576,7 +576,7 @@ export const Replicanti = {
       return player.replicanti.galaxies;
     },
     get extra() {
-      return Decimal.floor(new Decimal((new Decimal.timesEffectsOf(
+      return Decimal.floor(new Decimal((new Decimal(1).timesEffectsOf(
         TimeStudy(225),
         TimeStudy(226)
       ) + Effarig.bonusRG) * TimeStudy(303).effectOrDefault(1) * this.multiplication).add(this.bought.times(this.multiplication - 1)));
